chore(suite-native): adjust style for new design system
# Conflicts:
#	yarn.lock
