+++ suite-native/module-accounts-management/package.json
@@ -19,6 +19,7 @@
         "@react-navigation/native": "7.1.28",
         "@react-navigation/native-stack": "7.12.0",
         "@reduxjs/toolkit": "2.11.2",
+        "@suite-common/bip329": "workspace:*",
         "@suite-common/device": "workspace:*",
         "@suite-common/earn-stablecoin-api": "workspace:*",
         "@suite-common/fiat-services": "workspace:*",
@@ -54,6 +55,7 @@
         "@suite-native/services": "workspace:*",
         "@suite-native/tokens": "workspace:*",
         "@suite-native/transactions": "workspace:*",
+        "@trezor/connect": "workspace:*",
         "@trezor/styles-native": "workspace:*",
         "@trezor/utils": "workspace:*",
         "date-fns": "^4.1.0",
