+++ suite-native/module-accounts-management/src/screens/__tests__/AccountSettingsScreen.test.tsx
@@ -44,6 +44,7 @@ const buildPreloadedState = (account: ReturnType<typeof mockWalletAccount>) => (
     },
     suiteSyncData: { wallets: {} },
     wallet: { accounts: [account] },
+    suiteSync: { settings: {} },
 });
 
 describe('AccountSettingsScreen', () => {
